# Optimise mutations in single-pass epoch processing