--- a/consensus/state_processing/src/per_epoch_processing/single_pass.rs
+++ b/consensus/state_processing/src/per_epoch_processing/single_pass.rs
@@ -302,17 +302,17 @@ fn process_single_inactivity_update(
         if **inactivity_score == 0 {
             return Ok(());
         }
-        inactivity_score.make_mut().safe_sub_assign(1)?;
+        inactivity_score.make_mut()?.safe_sub_assign(1)?;
     } else {
         inactivity_score
-            .make_mut()
+            .make_mut()?
             .safe_add_assign(spec.inactivity_score_bias)?;
     }
 
     // Decrease the score of all validators for forgiveness when not during a leak
     if !state_ctxt.is_in_inactivity_leak {
         let deduction = min(spec.inactivity_score_recovery_rate, **inactivity_score);
-        inactivity_score.make_mut().safe_sub_assign(deduction)?;
+        inactivity_score.make_mut()?.safe_sub_assign(deduction)?;
     }
 
     Ok(())
@@ -349,7 +349,7 @@ fn process_single_reward_and_penalty(
     )?;
 
     if delta.rewards != 0 || delta.penalties != 0 {
-        let balance = balance.make_mut();
+        let balance = balance.make_mut()?;
         balance.safe_add_assign(delta.rewards)?;
         *balance = balance.saturating_sub(delta.penalties);
     }
@@ -462,7 +462,7 @@ fn process_single_registry_update(
     let current_epoch = state_ctxt.current_epoch;
 
     if validator.is_eligible_for_activation_queue(spec) {
-        validator.make_mut().mutable.activation_eligibility_epoch = current_epoch.safe_add(1)?;
+        validator.make_mut()?.mutable.activation_eligibility_epoch = current_epoch.safe_add(1)?;
     }
 
     if validator.is_active_at(current_epoch)
@@ -472,7 +472,7 @@ fn process_single_registry_update(
     }
 
     if activation_queue.contains(&validator_info.index) {
-        validator.make_mut().mutable.activation_epoch =
+        validator.make_mut()?.mutable.activation_epoch =
             spec.compute_activation_exit_epoch(current_epoch)?;
     }
 
@@ -509,7 +509,7 @@ fn initiate_validator_exit(
         exit_queue_epoch.safe_add_assign(1)?;
     }
 
-    let validator = validator.make_mut();
+    let validator = validator.make_mut()?;
     validator.mutable.exit_epoch = exit_queue_epoch;
     validator.mutable.withdrawable_epoch =
         exit_queue_epoch.safe_add(spec.min_validator_withdrawability_delay)?;
@@ -560,7 +560,7 @@ fn process_single_slashing(
             .safe_div(state_ctxt.total_active_balance)?
             .safe_mul(increment)?;
 
-        *balance.make_mut() = balance.saturating_sub(penalty);
+        *balance.make_mut()? = balance.saturating_sub(penalty);
     }
     Ok(())
 }
@@ -614,7 +614,7 @@ fn process_single_effective_balance_update(
     }
 
     if new_effective_balance != old_effective_balance {
-        validator.make_mut().mutable.effective_balance = new_effective_balance;
+        validator.make_mut()?.mutable.effective_balance = new_effective_balance;
 
         // Update progressive balances cache for the *current* epoch, which will soon become the
         // previous epoch once the epoch transition completes.