## Issue Addressed

Fix a `tree-states` performance regression introduced in #4483.

## Proposed Changes

Refactor single-pass epoch processing so that it avoids unnecessary mutations of the `validators`, `balances` and `inactivity_scores`. The initial version of single-pass epoch processing elected to mutate these values for every validator, even if no actual changes were required. Under the hood this caused the `milhouse` lists to allocate new nodes for all of the "updated" values, which wasted CPU time and memory. I think this could explain some of the memory issues observed in #4521.

The fix is to use `milhouse::Cow` (as we did previously), and only upgrade the cow to a mutable reference when a mutation _needs_ to be made.

Benchmarking `lcli transition-blocks` on an epoch boundary block showed a stark reduction in the time required to tree-hash the post-block state, from 210ms+ on my machine, down to 70ms (a 3x reduction).

## Additional Info

Waiting on https://github.com/sigp/milhouse/pull/22