@michaelsproul 2023-08-07
Benchmarks

Before:

```
[2023-08-07T05:15:04Z INFO  lcli::transition_blocks] Using mainnet spec
[2023-08-07T05:15:04Z INFO  lcli::transition_blocks] Doing 1 runs
[2023-08-07T05:15:04Z INFO  lcli::transition_blocks] Config { no_signature_verification: false, exclude_cache_builds: false, exclude_post_block_thc: false }
[2023-08-07T05:15:04Z INFO  lcli::transition_blocks] Block path: "/home/michael/Programming/lighthouse/block_6300864.ssz"
[2023-08-07T05:15:04Z INFO  lcli::transition_blocks] Pre-state path: "/home/michael/Programming/lighthouse/state_6300863.ssz"
[2023-08-07T05:15:04Z DEBUG lcli::transition_blocks] SSZ decoding /home/michael/Programming/lighthouse/state_6300863.ssz: 165.765329ms
[2023-08-07T05:15:04Z DEBUG lcli::transition_blocks] SSZ decoding /home/michael/Programming/lighthouse/block_6300864.ssz: 2.370627ms
[2023-08-07T05:15:04Z DEBUG lcli::transition_blocks] Building pubkey cache (might take some time)
[2023-08-07T05:15:34Z DEBUG lcli::transition_blocks] Build caches: 781.380927ms
[2023-08-07T05:15:34Z DEBUG lcli::transition_blocks] Initial tree hash: 46.345649ms
[2023-08-07T05:15:34Z DEBUG lcli::transition_blocks] Slot processing: 291.169759ms
[2023-08-07T05:15:34Z DEBUG lcli::transition_blocks] Build all caches (again): 870ns
[2023-08-07T05:15:34Z DEBUG lcli::transition_blocks] Batch verify block signatures: 9.287444ms
[2023-08-07T05:15:34Z DEBUG lcli::transition_blocks] Process block: 27.094788ms
[2023-08-07T05:15:34Z DEBUG lcli::transition_blocks] Post-block tree hash: 227.625198ms
[2023-08-07T05:15:34Z INFO  lcli::transition_blocks] Run 0: 1.383012308s
```

After:

```
[2023-08-07T05:58:04Z INFO  lcli::transition_blocks] Using mainnet spec
[2023-08-07T05:58:04Z INFO  lcli::transition_blocks] Doing 1 runs
[2023-08-07T05:58:04Z INFO  lcli::transition_blocks] Config { no_signature_verification: false, exclude_cache_builds: false, exclude_post_block_thc: false }
[2023-08-07T05:58:04Z INFO  lcli::transition_blocks] Block path: "/home/michael/Programming/lighthouse/block_6300864.ssz"
[2023-08-07T05:58:04Z INFO  lcli::transition_blocks] Pre-state path: "/home/michael/Programming/lighthouse/state_6300863.ssz"
[2023-08-07T05:58:04Z DEBUG lcli::transition_blocks] SSZ decoding /home/michael/Programming/lighthouse/state_6300863.ssz: 163.370615ms
[2023-08-07T05:58:04Z DEBUG lcli::transition_blocks] SSZ decoding /home/michael/Programming/lighthouse/block_6300864.ssz: 2.309347ms
[2023-08-07T05:58:04Z DEBUG lcli::transition_blocks] Building pubkey cache (might take some time)
[2023-08-07T05:58:33Z DEBUG lcli::transition_blocks] Build caches: 758.590088ms
[2023-08-07T05:58:33Z DEBUG lcli::transition_blocks] Initial tree hash: 44.091696ms
[2023-08-07T05:58:33Z DEBUG lcli::transition_blocks] Slot processing: 255.939054ms
[2023-08-07T05:58:33Z DEBUG lcli::transition_blocks] Build all caches (again): 950ns
[2023-08-07T05:58:33Z DEBUG lcli::transition_blocks] Batch verify block signatures: 8.519052ms
[2023-08-07T05:58:33Z DEBUG lcli::transition_blocks] Process block: 27.302409ms
[2023-08-07T05:58:33Z DEBUG lcli::transition_blocks] Post-block tree hash: 70.534357ms
[2023-08-07T05:58:33Z INFO  lcli::transition_blocks] Run 0: 1.165083305s
```